refactor after_category method in admin categories helper

+++ app/helpers/administration/categories_helper.rb
@@ -1,5 +1,4 @@
 module Administration::CategoriesHelper
-
   def unique_category_select_name(u)
     name = "#{u.company_name} (#{u.email})"
     name += " #{u.last_name}" if [:agent,:call_centre_agent].include?(u.role)
@@ -7,14 +6,17 @@ module Administration::CategoriesHelper
   end
 
   def after_category
-    lambda do |c|
-      [].tap do |links|
-          links << link_to(t("administration.categories.index.view.edit_link"), edit_administration_category_path(c))
-          links << link_to(t("administration.categories.index.view.destroy_link"), administration_category_path(c),
-                           method: :delete,
-                           confirm: I18n.t("common.confirmation")) if c.is_empty?
-      end.join(" | ").html_safe
+    ->(category) do
+      translation_scope = 'administration.categories.index.view'
+
+      links = []
+      links << link_to(t(:edit_link, scope: translation_scope), edit_administration_category_path(category))
+
+      if category.is_empty?
+        links << link_to(t(:destroy_link, scope: translation_scope), administration_category_path(category), method: :delete, confirm: I18n.t('common.confirmation'))
+      end
+
+      links.join(' | ').html_safe
     end
   end
-
 end
